Extract common expiry date validation logic to ui-core
COAND-683
